rename/product, category/#67 : folder rename vo->dto
