+++ src/main/java/com/ghbt/ghbt_starbucks/category/controller/CategoryController.java
@@ -1,9 +1,8 @@
 package com.ghbt.ghbt_starbucks.category.controller;
 
-import com.ghbt.ghbt_starbucks.category.model.Category;
 import com.ghbt.ghbt_starbucks.category.service.ICategoryService;
-import com.ghbt.ghbt_starbucks.category.vo.RequestCategory;
-import com.ghbt.ghbt_starbucks.category.vo.ResponseCategory;
+import com.ghbt.ghbt_starbucks.category.dto.RequestCategory;
+import com.ghbt.ghbt_starbucks.category.dto.ResponseCategory;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
