+++ src/main/java/com/ghbt/ghbt_starbucks/product/service/IProductService.java
@@ -3,8 +3,8 @@ package com.ghbt.ghbt_starbucks.product.service;
 import com.ghbt.ghbt_starbucks.product.Projection.IProductSearch;
 import com.ghbt.ghbt_starbucks.product.model.Product;
 import com.ghbt.ghbt_starbucks.product.Projection.IProductListByCategory;
-import com.ghbt.ghbt_starbucks.product.vo.RequestProduct;
-import com.ghbt.ghbt_starbucks.product.vo.ResponseProduct;
+import com.ghbt.ghbt_starbucks.product.dto.RequestProduct;
+import com.ghbt.ghbt_starbucks.product.dto.ResponseProduct;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
