+++ src/main/java/com/ghbt/ghbt_starbucks/product/service/ProductServiceImpl.java
@@ -7,8 +7,8 @@ import com.ghbt.ghbt_starbucks.product.Projection.IProductSearch;
 import com.ghbt.ghbt_starbucks.product.model.Product;
 import com.ghbt.ghbt_starbucks.product.Projection.IProductListByCategory;
 import com.ghbt.ghbt_starbucks.product.repository.IProductRepository;
-import com.ghbt.ghbt_starbucks.product.vo.RequestProduct;
-import com.ghbt.ghbt_starbucks.product.vo.ResponseProduct;
+import com.ghbt.ghbt_starbucks.product.dto.RequestProduct;
+import com.ghbt.ghbt_starbucks.product.dto.ResponseProduct;
 import com.ghbt.ghbt_starbucks.product_and_category.model.ProductAndCategory;
 import com.ghbt.ghbt_starbucks.product_and_category.repository.IProductAndCategoryRepository;
 import lombok.Data;
