+++ src/main/java/com/ghbt/ghbt_starbucks/product_and_category/dto/ResponseProductAndCategory.java
@@ -1,4 +1,4 @@
-package com.ghbt.ghbt_starbucks.product_and_category.vo;
+package com.ghbt.ghbt_starbucks.product_and_category.dto;
 
 import com.ghbt.ghbt_starbucks.category.model.Category;
 import com.ghbt.ghbt_starbucks.product.model.Product;
